requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.20
psycopg[binary]==3.2.3

# LangChain & ecosystem
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.79
langchain-text-splitters==0.3.10
langchain-huggingface==0.3.1
langchain-openai==0.3.32
openai==1.102.0
chromadb==0.5.5
sentence-transformers==5.1.2

# Document loaders
pypdf==4.2.0
python-docx==1.2.0

# Utils
tqdm==4.67.1
SQLAlchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.1
python-json-logger==2.0.7
sentry-sdk==2.19.2
slowapi==0.1.9
redis==5.2.1
huggingface-hub>=0.30.0
transformers>=4.30.0
torch>=2.0.0

